Extractd contacts class
